Move metrics to a separate package
